requirements.txt
streamlit
google-generativeai
protobuf